+++ src/Apps/Order/Utils/__tests__/formValidators.jest.tsx
@@ -1,4 +1,4 @@
-import { Address } from "Components/Address/AddressForm"
+import { Address } from "Components/Address/LegacyAddressForm"
 import {
   validateAddress,
   validatePresence,
