+++ src/Apps/Order/Utils/formValidators.tsx
@@ -1,4 +1,4 @@
-import { Address } from "Components/Address/AddressForm"
+import { Address } from "Components/Address/LegacyAddressForm"
 import { isEmpty } from "lodash"
 
 export const validatePresence = (value: string): string | null => {
